Move notifications module to user_interface subpackage
